Clarify allowed format for repositories key
Using JSON object should be discouraged as there is no explicit order
of keys in JSON object. Even though it's deterministic in current PHP
version, it may change any time. As stated on http://json.org/:
"An object is an unordered set of name/value pairs. An object begins
with { (left brace) and ends with } (right brace). Each name is
followed by : (colon) and the name/value pairs are separated
by , (comma)."

Fixes #2802.